requirements.txt
streamlit>=1.28.0
streamlit-authenticator>=0.3.2
PyYAML>=6.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
extra-streamlit-components>=0.1.60